Safari-611.1.10/LayoutTests/fast/dom/Element/documentElement-scrollLeft-Quirks.html

<html>
    <head>
        <style>
            div {
                height: 9999px;
                width:9999px;
            }
        </style>
        <script src="../../../resources/js-test-pre.js"></script>
        <script>
            function runTest() {
                description('Tests that for non-standard mode document.documentElement.scrollLeft returns the scroll left value as 0');

                document.documentElement.scrollLeft = 500;
                shouldBe("document.documentElement.scrollLeft","0");
				isSuccessfullyParsed();
            }
        </script>
    </head>
    <body onload="runTest()">
        <div></div>
        <script src="../../../resources/js-test-post.js"></script>
    </body>
</html>
