LayoutTests/fast/dom/Element/documentElement-scrollLeft-Quirks-expected.txt

Tests that for non-standard mode document.documentElement.scrollLeft returns the scroll left value as 0

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS document.documentElement.scrollLeft is 0
PASS successfullyParsed is true

TEST COMPLETE
